ptclient: fix various compile warnings
Some due to old code, some due to partially-applied API changes in the
2.5 series.

+++ ptclient/ldap.c
@@ -351,7 +351,8 @@ static int ptsmodule_connect(void)
 
 	rc = ldap_set_option(ptsm->ld, LDAP_OPT_NETWORK_TIMEOUT, &(ptsm->timeout));
 	if (rc != LDAP_OPT_SUCCESS) {
-		syslog(LOG_WARNING, "Unable to set LDAP_OPT_NETWORK_TIMEOUT %d.%d.", ptsm->timeout.tv_sec, ptsm->timeout.tv_usec);
+		syslog(LOG_WARNING, "Unable to set LDAP_OPT_NETWORK_TIMEOUT %ld.%06d.",
+		       (long)ptsm->timeout.tv_sec, (int)ptsm->timeout.tv_usec);
 	}
 
 	rc = ldap_set_option(ptsm->ld, LDAP_OPT_TIMELIMIT, &(ptsm->time_limit));
